Fix compatibility issues for PlayImport
The emojiLogs for 1.0 is now just printing a message
about not being supported anymore. We can later
completely remove it.
