+++ framework/src/sbt-plugin/src/main/scala-sbt-1.0/play/sbt/PlayImportCompat.scala
@@ -1,0 +1,23 @@
+/*
+ * Copyright (C) 2009-2017 Lightbend Inc. <https://www.lightbend.com>
+ */
+package play.sbt
+
+import sbt.Keys._
+import sbt.{ Def, Level, Scope, Settings, State }
+import sbt.internal.LogManager
+import sbt.internal.util.ManagedLogger
+
+/**
+ * Fix compatibility issues for PlayImport. This is the version compatible with sbt 1.0.
+ */
+trait PlayImportCompat {
+
+  /**
+   * This is not supported anymore.
+   */
+  def emojiLogs = {
+    sLog.value.warn("emojiLogs is not supported anymore. Remove it from your build.")
+  }
+
+}
