+++ framework/src/sbt-plugin/src/main/scala/play/sbt/PlayImport.scala
@@ -3,7 +3,6 @@
  */
 package play.sbt
 
-import sbt.Keys._
 import sbt._
 
 import play.dev.filewatch.FileWatchService
@@ -11,7 +10,7 @@ import play.dev.filewatch.FileWatchService
 /**
  * Declares the default imports for Play plugins.
  */
-object PlayImport {
+object PlayImport extends PlayImportCompat {
 
   val Production = config("production")
 
@@ -71,39 +70,6 @@ object PlayImport {
   val openId = component("play-openid")
 
   val specs2 = component("play-specs2")
-
-  /**
-   * Add this to your build.sbt, eg:
-   *
-   * {{{
-   *   emojiLogs
-   * }}}
-   */
-  lazy val emojiLogs = logManager ~= { lm =>
-    new LogManager {
-      def apply(data: Settings[Scope], state: State, task: Def.ScopedKey[_], writer: java.io.PrintWriter) = {
-        val l = lm.apply(data, state, task, writer)
-        val FailuresErrors = "(?s).*(\\d+) failures?, (\\d+) errors?.*".r
-        new Logger {
-          def filter(s: String) = {
-            val filtered = s.replace("\033[32m+\033[0m", "\u2705 ")
-              .replace("\033[33mx\033[0m", "\u274C ")
-              .replace("\033[31m!\033[0m", "\uD83D\uDCA5 ")
-            filtered match {
-              case FailuresErrors("0", "0") => filtered + " \uD83D\uDE04"
-              case FailuresErrors(_, _) => filtered + " \uD83D\uDE22"
-              case _ => filtered
-            }
-          }
-          def log(level: Level.Value, message: => String) = l.log(level, filter(message))
-          def success(message: => String) = l.success(message)
-          def trace(t: => Throwable) = l.trace(t)
-
-          override def ansiCodesSupported = l.ansiCodesSupported
-        }
-      }
-    }
-  }
 
   object PlayKeys {
     val playDefaultPort = SettingKey[Int]("playDefaultPort", "The default port that Play runs on")
